requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.3
django-filter==24.2
django-soft-delete==1.0.12
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.3.4
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
pycryptodome==3.20.0
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.34.0
rest-framework-simplejwt==0.0.2
rpds-py==0.18.0
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
